--- a/addons/logistics_wirecutter/README.md
+++ b/addons/logistics_wirecutter/README.md
@@ -1,7 +1,7 @@
 ace_logistics_wirecutter
 ===========
 
-Adds an item that allows cutting of fences in Aarma 3 and AiA/CUP maps.
+Adds an item that allows cutting of fences in Arma 3 and AiA/CUP maps.
 
 #### Items Added:
 `ACE_wirecutter`