--- a/addons/logistics_wirecutter/XEH_PREP.hpp
+++ b/addons/logistics_wirecutter/XEH_PREP.hpp
@@ -1,4 +1,3 @@
-
 PREP(cutDownFence);
 PREP(interactEH);
 PREP(isFence);