--- a/addons/logistics_wirecutter/data/ace_wirecutter.rvmat
+++ b/addons/logistics_wirecutter/data/ace_wirecutter.rvmat
@@ -1,6 +1,5 @@
-class StageTI
-{
-        texture = "a3\data_f\default_glass_ti_ca.paa";
+class StageTI {
+    texture = "a3\data_f\default_glass_ti_ca.paa";
 };
 ambient[] = {1,1,1,1};
 diffuse[] = {1,1,1,1};
@@ -10,81 +9,71 @@ specular[] = {0,0,0,1};
 specularPower = 0;
 PixelShaderID = "Super";
 VertexShaderID = "Super";
-class Stage1
-{
-        texture = "z\ace\addons\logistics_wirecutter\data\ace_wirecutter_nohq.paa";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1,0,0};
-                up[] = {0,1,0};
-                dir[] = {0,0,0};
-                pos[] = {0,0,0};
-        };
+class Stage1 {
+    texture = "z\ace\addons\logistics_wirecutter\data\ace_wirecutter_nohq.paa";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1,0,0};
+        up[] = {0,1,0};
+        dir[] = {0,0,0};
+        pos[] = {0,0,0};
+    };
 };
-class Stage2
-{
-        texture = "#(argb,8,8,3)color(0.5,0.5,0.5,0.5,DT)";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1,0,0};
-                up[] = {0,1,0};
-                dir[] = {0,0,0};
-                pos[] = {0,0,0};
-        };
+class Stage2 {
+    texture = "#(argb,8,8,3)color(0.5,0.5,0.5,0.5,DT)";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1,0,0};
+        up[] = {0,1,0};
+        dir[] = {0,0,0};
+        pos[] = {0,0,0};
+    };
 };
 class Stage3
 {
-        texture = "#(argb,8,8,3)color(0,0,0,0,MC)";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1,0,0};
-                up[] = {0,1,0};
-                dir[] = {0,0,0};
-                pos[] = {0,0,0};
-        };
+    texture = "#(argb,8,8,3)color(0,0,0,0,MC)";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1,0,0};
+        up[] = {0,1,0};
+        dir[] = {0,0,0};
+        pos[] = {0,0,0};
+    };
 };
 class Stage4
 {
-        texture = "#(argb,8,8,3)color(1,1,1,1,AS)";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1,0,0};
-                up[] = {0,1,0};
-                dir[] = {0,0,0};
-                pos[] = {0,0,0};
-        };
+    texture = "#(argb,8,8,3)color(1,1,1,1,AS)";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1,0,0};
+        up[] = {0,1,0};
+        dir[] = {0,0,0};
+        pos[] = {0,0,0};
+    };
 };
 class Stage5
 {
-        texture = "#(argb,8,8,3)color(0,0.6,1,1,SMDI)";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1,0,0};
-                up[] = {0,1,0};
-                dir[] = {0,0,0};
-                pos[] = {0,0,0};
-        };
+    texture = "#(argb,8,8,3)color(0,0.6,1,1,SMDI)";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1,0,0};
+        up[] = {0,1,0};
+        dir[] = {0,0,0};
+        pos[] = {0,0,0};
+    };
 };
-class Stage6
-{
-        texture = "#(ai,32,128,1)fresnel(0.71,0.74)";
-        uvSource = "none";
+class Stage6 {
+    texture = "#(ai,32,128,1)fresnel(0.71,0.74)";
+    uvSource = "none";
 };
-class Stage7
-{
-        texture = "a3\data_f\env_land_co.paa";
-        useWorldEnvMap = "true";
-        uvSource = "tex";
-        class uvTransform
-        {
-                aside[] = {1.0,0.0,0.0};
-                up[] = {0.0,1.0,0.0};
-                dir[] = {0.0,0.0,0.0};
-                pos[] = {0.0,0.0,0.0};
-        };
+class Stage7 {
+    texture = "a3\data_f\env_land_co.paa";
+    useWorldEnvMap = "true";
+    uvSource = "tex";
+    class uvTransform {
+        aside[] = {1.0,0.0,0.0};
+        up[] = {0.0,1.0,0.0};
+        dir[] = {0.0,0.0,0.0};
+        pos[] = {0.0,0.0,0.0};
+    };
 };