--- a/addons/logistics_wirecutter/functions/fnc_cutDownFence.sqf
+++ b/addons/logistics_wirecutter/functions/fnc_cutDownFence.sqf
@@ -1,6 +1,6 @@
 /*
- * Author: gpgpgpgp, edited by commy2, PabstMirror
- * Starts cutting down a fence
+ * Author: gpgpgpgp, commy2, PabstMirror, mharis001
+ * Starts cutting down a fence. Triggers global "ace_wireCuttingStarted" event.
  *
  * Arguments:
  * 0: Unit <OBJECT>
@@ -10,52 +10,60 @@
  * None
  *
  * Example:
- * [player, berlinWall] call ace_logistics_wirecutter_fnc_cutDownFence
+ * [_player, _fence] call ace_logistics_wirecutter_fnc_cutDownFence
  *
  * Public: No
  */
 #include "script_component.hpp"
 
-params ["_unit", "_fenceObject"];
-TRACE_2("params",_unit,_fenceObject);
+params ["_unit", "_fence"];
+TRACE_2("Fence cutting started",_unit,_fence);
 
+// Exit if unit is not player
 if (_unit != ACE_player) exitWith {};
 
-private _timeToCut = if ([_unit] call EFUNC(common,isEngineer)) then {7.5} else {11};
+// Get cut time based on if unit is a engineer
+private _cutTime = if (_unit call EFUNC(common,isEngineer)) then {CUT_TIME_ENGINEER} else {CUT_TIME_DEFAULT};
 
 if !(_unit call EFUNC(common,isSwimming)) then {
     [_unit, "AinvPknlMstpSnonWnonDr_medic5", 0] call EFUNC(common,doAnimation);
 };
 
-private _onCompletion = {
-    TRACE_1("_onCompletion",_this);
-    (_this select 0) params ["_fenceObject", "", "_unit"];
-    _fenceObject setdamage 1;
-    if !(_unit call EFUNC(common,isSwimming)) then {
-        [_unit, "AmovPknlMstpSrasWrflDnon", 1] call EFUNC(common,doAnimation);
-    };
-};
+[
+    _cutTime,
+    [_unit, _fence, 0],
+    {
+        TRACE_1("Fence cutting successful",_this);
+        (_this select 0) params ["_unit", "_fence"];
 
-private _onFail = {
-    TRACE_1("_onFail", _this);
-    (_this select 0) params ["", "", "_unit"];
-    if !(_unit call EFUNC(common,isSwimming)) then {
-        [_unit, "AmovPknlMstpSrasWrflDnon", 1] call EFUNC(common,doAnimation);
-    };
-};
+        _fence setDamage 1;
+        if !(_unit call EFUNC(common,isSwimming)) then {
+            [_unit, "AmovPknlMstpSrasWrflDnon", 1] call EFUNC(common,doAnimation);
+        };
+    },
+    {
+        TRACE_1("Fence cutting failed",_this);
+        (_this select 0) params ["_unit"];
 
-private _progressCheck = {
-    params ["_args", "_passedTime"];
-    _args params ["_fenceObject", "_lastSoundEffectTime", "_unit"];
+        if !(_unit call EFUNC(common,isSwimming)) then {
+            [_unit, "AmovPknlMstpSrasWrflDnon", 1] call EFUNC(common,doAnimation);
+        };
+    },
+    localize LSTRING(CuttingFence),
+    {
+        params ["_args", "_elapsedTime"];
+        _args params ["_unit", "_fence", "_lastSoundTime"];
 
-    if (_passedTime > (_lastSoundEffectTime + SOUND_CLIP_TIME_SPACEING)) then {
-        playSound3D [QUOTE(PATHTO_R(sound\wirecut.ogg)), objNull, false, (getPosASL _unit), 3, 1, 10];
-        _args set [1, _passedTime];
-    };
+        if (_elapsedTime > _lastSoundTime + SOUND_CLIP_TIME_SPACING) then {
+            playSound3D [QPATHTO_R(sound\wirecut.ogg), objNull, false, getPosASL _unit, 3, 1, 10];
+            _args set [2, _elapsedTime];
+        };
 
-    ((!isNull _fenceObject) && {(damage _fenceObject) < 1} && {("ACE_wirecutter" in (items _unit))})
-};
+        !isNull _fence
+        && {damage _fence < 1}
+        && {"ACE_wirecutter" in ([_player, false, true, true, true, false] call CBA_fnc_uniqueUnitItems)}
+    },
+    ["isNotSwimming"]
+] call EFUNC(common,progressBar);
 
-[_timeToCut, [_fenceObject,0,_unit], _onCompletion, _onFail, localize LSTRING(CuttingFence), _progressCheck, ["isNotSwimming"]] call EFUNC(common,progressBar);
-
-["ace_wireCuttingStarted", [_unit, _fenceObject]] call CBA_fnc_globalEvent;
+["ace_wireCuttingStarted", [_unit, _fence]] call CBA_fnc_globalEvent;