--- a/addons/logistics_wirecutter/functions/fnc_isFence.sqf
+++ b/addons/logistics_wirecutter/functions/fnc_isFence.sqf
@@ -1,16 +1,16 @@
 /*
  * Author: PabstMirror
- * Checks if object is a fence.  Should work on any fence type, even (typeOf == "").
- * Call is fairly expensive because of all of the string checking.
+ * Checks if object is a fence. Should work on any fence type, even when (typeOf == "").
+ * Call is fairly expensive because of string checking.
  *
  * Arguments:
- * 0: An Object To Test <OBJECT>
+ * 0: Object to test <OBJECT>
  *
  * Return Value:
- * Is it a fence <BOOL>
+ * Is fence <BOOL>
  *
  * Example:
- * [aFence] call ace_logistics_wirecutter_fnc_isFence
+ * [_fence] call ace_logistics_wirecutter_fnc_isFence
  *
  * Public: No
  */
@@ -22,11 +22,11 @@ TRACE_1("params",_object);
 private _typeOf = typeOf _object;
 
 private _returnValue = if (_typeOf != "") then {
-    //If the fence has configEntry we can check it directly
-    (1 == (getNumber (configFile >> "CfgVehicles" >> _typeOf >> QGVAR(isFence))));
+    // Check for isFence entry since we have valid typeOf
+    1 == getNumber (configFile >> "CfgVehicles" >> _typeOf >> QGVAR(isFence));
 } else {
-    //Check the p3d name against list (in script_component.hpp)
-    ((getModelInfo _object) select 0) in FENCE_P3DS;
+    // Check the p3d name against list (in script_component.hpp)
+    (getModelInfo _object select 0) in FENCE_P3DS;
 };
 
 _returnValue