--- a/addons/logistics_wirecutter/script_component.hpp
+++ b/addons/logistics_wirecutter/script_component.hpp
@@ -16,8 +16,7 @@
 
 #include "\z\ace\addons\main\script_macros.hpp"
 
-
-//find is case sensitive, so keep everything lowercase
+// find is case sensitive, so keep everything lowercase
 #define FENCE_P3DS [\
     "mil_wiredfence_f.p3d",\
     "wall_indfnc_3.p3d",\
@@ -26,7 +25,19 @@
     "pletivo_wired.p3d",\
     "wall_fen1_5.p3d",\
     "net_fence_8m_f.p3d",\
-    "razorwire_f.p3d"\
+    "razorwire_f.p3d",\
+    "wired_fence_4m_f.p3d",\
+    "wired_fence_8m_f.p3d",\
+    "plasticnetfence_01_short_f.p3d",\
+    "plasticnetfence_01_long_f.p3d",\
+    "netfence_01_m_4m_f.p3d",\
+    "netfence_01_m_8m_f.p3d",\
+    "wiredfence_01_4m_f.p3d",\
+    "wiredfence_01_8m_f.p3d",\
+    "wiredfence_01_16m_f.p3d",\
+    "slums01_8m.p3d",\
+    "slums02_4m.p3d",\
+    "backalley_01_l_1m_f.p3d"\
 ]
 
-#define SOUND_CLIP_TIME_SPACEING    1.5
+#define SOUND_CLIP_TIME_SPACEING 1.5