--- a/addons/logistics_wirecutter/script_component.hpp
+++ b/addons/logistics_wirecutter/script_component.hpp
@@ -40,4 +40,6 @@
     "backalley_01_l_1m_f.p3d"\
 ]
 
-#define SOUND_CLIP_TIME_SPACEING 1.5
+#define SOUND_CLIP_TIME_SPACING 1.5
+#define CUT_TIME_DEFAULT 11
+#define CUT_TIME_ENGINEER 7.5