… addons/logistics_wirecutter/CfgSounds.hpp
… addons/logistics_wirecutter/sound/wire_cut.ogg
… addons/logistics_wirecutter/sound/wire_cut_long.ogg