**When merged this pull request will:**
- Optimize `interactEH` function
- Misc improvements to `isFence` and `cutDownFence` functions (cleanup, readability, time macros)
- Remove redundant parentheses and spaces
- Added missing fences
- RVMAT cleanup, config guidelines
- Remove unused sound files

Performance results for `ace_logistics_wirecutter_interactEH_counter`:

Test | Old | New
--- | --- | ---
1 | 0.320984s / 433 = 0.741302ms | 0.169006s / 415 = 0.407244ms
2 | 0.307983s / 290 = 1.06201ms | 0.128906s / 334 = 0.385947ms
3 | 0.611984s / 419 = 1.46058ms | 0.183029s / 425 = 0.430657ms