@mharis001 2018-06-10
Also, currently an action can be triggered from 5.5m away. This is probably done to show the action from far for long fence objects. The statement could do a separate check for a smaller distance like 2m and if the player is farther away, a message could be shown.