@mharis001 2018-06-10
They sound like a machine cutting metal (a saw or something similar), instead of clipping with a wire cutter.